+++ assets/styles/_base.css
@@ -8,19 +8,17 @@
 :root {
   --color-not-white: #fafbfc;
   --color-light-gray: #efeef3;
+  --color-light-gray-see-thru: #efeef355;
   --color-gray: #dfe3e7;
   --color-dark-gray: #252228;
   --color-not-black: #09090C;
 }
 :root {
-  --background: var(--color-not-black);
-  --text: var(--color-not-white);
-  --nav-active-background: var(--color-dark-gray);
-}
-:root.light-mode {
   --background: var(--color-light-gray);
+  --background-see-thru: var(--color-light-gray-see-thru);
   --nav-active-background: white;
   --text: var(--color-not-black);
+  --thumb-size: clamp(4em, 5.5em, 15vw);
 }
 
 html {
@@ -40,24 +38,21 @@ body {
   overflow-x: hidden;
 }
 
-main {
-  flex-shrink: 0;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 50em;
-  margin: 0 auto;
-  padding: 2em 2em 1em 2em;
-}
+/* Components */
 
-nav {
+.nav {
   display: flex;
   align-items: center;
   align-self: stretch;
   padding: 1.1em;
   overflow-y: auto;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 1;
+  width: 100%;
 }
-nav a {
+.nav .link {
   font: inherit;
   color: inherit;
   text-decoration: none;
@@ -75,16 +70,33 @@ nav a {
   overflow: hidden;
   flex-shrink: 0;
 }
-nav a:first-child {
+.nav .link:first-child {
   margin-left: auto;
 }
-nav a:last-child {
+.nav .link:last-child {
   margin-right: auto;
 }
-nav a:hover,
-nav a.current {
+.nav .link:hover,
+.nav .link.active {
   background: var(--nav-active-background);
   opacity: 1;
+  box-shadow: 0 0.5em 0.5em #00000005;
+}
+
+.construction {
+  width: 100vw;
+  height: 100vh;
+  font-size: clamp(1.5em, 3em, 4vw);
+  color: #000000ff;
+  padding: 1em;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: 900;
+  letter-spacing: 0.03em;
+  word-spacing: 0.2em;
+  text-transform: uppercase;
+  opacity: 0.3;
 }
 
 /* Utilities */
