+++ assets/styles/posts.css
@@ -1,11 +1,3 @@
-:root {
-  --thumb-size: clamp(4em, 5.5em, 15vw);
-}
-
-main {
-  padding: 2.5em 2.75em 0 2.5em;
-}
-
 .header {
   align-self: center;
   display: grid;
@@ -14,7 +6,7 @@ main {
   grid-template-rows: 0.65em 0.35em;
   align-items: center;
   gap: 0.1em 0.4em;
-  margin: 0.1em 0 0.75em 0;
+  margin: 7.5rem 0 0.8em 0;
 }
 .header .name {
   font-size: 0.5em;
@@ -48,9 +40,22 @@ main {
   font-size: 0.10em;
 }
 
-.article {
+.main {
+  flex-shrink: 0;
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  width: 100%;
+  max-width: 50em;
+  /* Extra right padding to accomadate the thumbnails */
+  padding: 0 2.75em 2.5em 2.5em;
+}
+
+.post {
+  /* FIXME: I shouldn't have done this first font-size. The layout is near
+  perfect but replicating it based around --thumb-size is now difficult because
+  I coupled it with this font-size from the get-go */
   font-size: clamp(0.8em, 1em, 3vw);
-  display: block;
   flex-shrink: 0;
   text-rendering: optimizeLegibility;
   background: white;
@@ -63,9 +68,9 @@ main {
   box-shadow: 0 1em 2em #0000000B;
   color: inherit;
   text-decoration: none;
-  transition: background 0.15s;
+  display: block;
 }
-.article .thumb {
+.post .thumb {
   font-size: var(--thumb-size);
   width: 1em;
   height: 1em;
@@ -73,27 +78,30 @@ main {
   top: -0.3em;
   right: -0.15em;
   border-radius: 0.3em;
-  z-index: 2;
-}
-.article .thumb.gif {
   z-index: 1;
+  color: transparent;
+  user-select: none;
 }
-.article:hover .thumb.gif {
-  z-index: 3;
+.post .thumb.gif {
+  z-index: 2;
+  visibility: hidden;
 }
-.article .title {
+.post:hover .thumb.gif {
+  visibility: visible;
+}
+.post .title {
   font-size: calc(0.25 * var(--thumb-size));
   margin-bottom: 0.75em;
   margin-right: 2.5em;
   font-weight: 900;
   letter-spacing: 0.02em;
 }
-.article .description {
+.post .description {
   font-size: calc(0.2 * var(--thumb-size));
   line-height: 1.5;
   opacity: 0.6;
 }
-.article .date {
+.post .date {
   position: absolute;
   right: 1.25em;
   bottom: -2em;
