+++ pages/mod.ts
@@ -1,0 +1,6 @@
+
+export * from "./about.ts";
+export * from "./article.ts";
+export * from "./posts.ts";
+export * from "./projects.ts";
+export * from "./soc.ts";
